Replace use of DBPedia with the new SimpleHTTPMock
This should eliminate the spurious DBPedia related errors that plagued
the CI pipeline.

I am deleting `test/test_core_sparqlstore.py` as it does not do anything
meaningful as far as I can tell and certainly nothing different from
what is being done in `test/test_sparqlstore.py`.

Also adds ServedSimpleHTTPMock to testutils.py which inherits from
SimpleHTTPMock and adds a http server to it which makes it a bit less
fiddly to use in tests.