Narrow absolute URL check and add some tests for norm_url
There are strings that contain `://` that are not valid absolute URLs,
for example, `://` itself is not an absolute URL, and something like
`a/b/c/d/e://` would also not be an asbolute URL.

This change changes the absolute URL check to only match actual absolute
URLs.

Also add tests for norm_url.